Prepare store refactoring by renaming old stores folder
