+++ libs/search-widget/src/core/old-stores/annotation.store.ts
