+++ libs/search-widget/src/core/old-stores/main.store.ts
