+++ libs/search-widget/src/core/old-stores/resource.store.ts
