+++ libs/search-widget/src/core/old-stores/viewer.store.ts
