+++ libs/search-widget/src/old-components/embedded-search/EmbeddedSearch.spec.ts
@@ -1,6 +1,6 @@
 import EmbeddedSearch from './EmbeddedSearch.svelte';
 import { fireEvent, render } from '@testing-library/svelte';
-import { nucliaStore } from '../../core/stores/main.store';
+import { nucliaStore } from '../../core/old-stores/main.store';
 import { firstValueFrom } from 'rxjs';
 import type { IResource } from '@nuclia/core';
