+++ libs/search-widget/src/old-components/viewer/InputViewer.svelte
@@ -2,7 +2,7 @@
   import { _ } from '../../core/i18n';
   import { getCDN } from '../../core/utils';
   import { SearchOrder } from '../../core/models';
-  import { viewerStore, viewerState, clearSearch } from '../../core/stores/viewer.store';
+  import { viewerStore, viewerState, clearSearch } from '../../core/old-stores/viewer.store';
 
   const query = viewerStore.query;
   query['set'] = query.next;
