+++ libs/search-widget/src/old-components/viewer/menus/EntityFamilyMenu.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { nucliaState } from '../../../core/stores/main.store';
+  import { nucliaState } from '../../../core/old-stores/main.store';
   import { clickOutside } from '../../../common/actions/actions';
   import { createEventDispatcher, onMount } from 'svelte';
   import { _ } from '../../../core/i18n';
