+++ libs/search-widget/src/old-components/viewer/menus/LabelMenu.svelte
@@ -2,8 +2,8 @@
   import { createEventDispatcher } from 'svelte';
   import { map } from 'rxjs';
   import { getCDN } from '../../../core/utils';
-  import { nucliaState } from '../../../core/stores/main.store';
-  import { viewerStore } from '../../../core/stores/viewer.store';
+  import { nucliaState } from '../../../core/old-stores/main.store';
+  import { viewerStore } from '../../../core/old-stores/viewer.store';
   import { clickOutside } from '../../../common/actions/actions';
   import Label from '../../../common/label/Label.svelte';
   import type { ParagraphLabels } from '../../../core/models';
