+++ libs/search-widget/src/widgets/search-widget/Widget.svelte
@@ -3,7 +3,7 @@
 <script lang="ts">
   import PopupSearch from '../../old-components/popup-search/PopupSearch.svelte';
   import EmbeddedSearch from '../../old-components/embedded-search/EmbeddedSearch.svelte';
-  import { nucliaState, setWidgetActions, resetStore, setDisplayedResource } from '../../core/stores/main.store';
+  import { nucliaState, setWidgetActions, resetStore, setDisplayedResource } from '../../core/old-stores/main.store';
   import { getResource, initNuclia, resetNuclia } from '../../core/api';
   import { concatMap, filter, tap } from 'rxjs/operators';
   import { onMount } from 'svelte';
